FFL-1016 Add FlagsClientProtocol
