+++ DatadogFlags/Sources/Client/FlagsClient.swift
@@ -69,7 +69,9 @@ public class FlagsClient {
             featureScope: featureScope
         )
     }
+}
 
+extension FlagsClient: FlagsClientProtocol {
     public func setEvaluationContext(_ context: FlagsEvaluationContext, completion: @escaping (Result<Void, FlagsError>) -> Void) {
         featureScope.context { [httpClient, configuration] sdkContext in
             httpClient.postPrecomputeAssignments(
@@ -132,66 +134,5 @@ public class FlagsClient {
         }
 
         return details
-    }
-}
-
-// MARK: - Convenience flag evaluation methods
-
-extension FlagsClient {
-    @inlinable
-    public func getValue<T>(key: String, defaultValue: T) -> T where T: Equatable, T: FlagValue {
-        getDetails(key: key, defaultValue: defaultValue).value
-    }
-
-    @inlinable
-    public func getBooleanValue(key: String, defaultValue: Bool) -> Bool {
-        getValue(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getStringValue(key: String, defaultValue: String) -> String {
-        getValue(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getIntegerValue(key: String, defaultValue: Int) -> Int {
-        getValue(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getDoubleValue(key: String, defaultValue: Double) -> Double {
-        getValue(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getObjectValue(key: String, defaultValue: AnyValue) -> AnyValue {
-        getValue(key: key, defaultValue: defaultValue)
-    }
-}
-
-extension FlagsClient {
-    @inlinable
-    public func getBooleanDetails(key: String, defaultValue: Bool) -> FlagDetails<Bool> {
-        getDetails(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getStringDetails(key: String, defaultValue: String) -> FlagDetails<String> {
-        getDetails(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getIntegerDetails(key: String, defaultValue: Int) -> FlagDetails<Int> {
-        getDetails(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getDoubleDetails(key: String, defaultValue: Double) -> FlagDetails<Double> {
-        getDetails(key: key, defaultValue: defaultValue)
-    }
-
-    @inlinable
-    public func getObjectDetails(key: String, defaultValue: AnyValue) -> FlagDetails<AnyValue> {
-        getDetails(key: key, defaultValue: defaultValue)
     }
 }
